Add an rbuild option.

+++ scripts/build_migraphx.sh
@@ -1,24 +1,29 @@
 #!/bin/bash
 #
 # build migraphx in docker container
-env HIP_DOCKER_WORKAROUND=${HIP_DOCKER_WORKAROUND:="0"}
+env USE_RBUILD=${USE_RBUILD:="0"}
 cd /src/AMDMIGraphX
-if [ -d build ]; then
-    rm -rf build
-fi
-mkdir build
-cd build
-# workaround for hip packaging in docker
-if [ -f '/opt/rocm/llvm/bin/clang++' ]; then
-    env CXX=/opt/rocm/llvm/bin/clang++ CXXFLAGS="-O3" cmake ..    
-elif [ "$HIP_DOCKER_WORKAROUND" = "1" ]; then
-    env CXX=/opt/rocm/bin/hcc CXXFLAGS="-O3 -fno-gpu-rdc -amdgpu-target=gfx906" cmake ..
+if [ "$USE_RBUILD" = "0" ]; then
+    if [ -d build ]; then
+	rm -rf build
+    fi
+    mkdir build
+    cd build
+    # workaround for hip packaging in docker
+    if [ -f '/opt/rocm/llvm/bin/clang++' ]; then
+	env CXX=/opt/rocm/llvm/bin/clang++ CXXFLAGS="-O3" cmake ..    
+    else
+	echo "Missing clang++"
+	exit 1
+    fi
+
+    if [ "$LD_LIBRARY_PATH" = "" ]; then
+	export LD_LIBRARY_PATH=/usr/local/lib:
+    fi
+
+    make -j4
 else
-    env CXX=/opt/rocm/bin/hcc CXXFLAGS="-O3 -fno-gpu-rdc" cmake ..
+    pip3 install https://github.com/RadeonOpenCompute/rbuild/archive/master.tar.gz
+    rbuild build -d depend -B build --cxx=/opt/rocm/llvm/bin/clang++
 fi
 
-if [ "$LD_LIBRARY_PATH" = "" ]; then
-    export LD_LIBRARY_PATH=/usr/local/lib:
-fi
-
-make -j4
